Move session and subscription request types into metabase-types/api
LoginData moves out of metabase/redux/auth and DraftDashboardSubscription /
DashboardSubscriptionData move out of metabase/redux/store, so metabase/api
no longer imports these types from redux above it.
