+++ frontend/src/metabase-types/api/session.ts
@@ -1,3 +1,9 @@
 export interface PasswordResetTokenStatus {
   valid: boolean;
 }
+
+export interface LoginData {
+  username: string;
+  password: string;
+  remember?: boolean;
+}
