+++ frontend/src/metabase/dashboard/components/DashboardSubscriptionsSidebar/DashboardSubscriptionsSidebar.tsx
@@ -22,7 +22,7 @@ import {
 } from "metabase/notifications/pulse/selectors";
 import { NEW_PULSE_TEMPLATE, cleanPulse, createChannel } from "metabase/pulse";
 import { connect, useDispatch } from "metabase/redux";
-import type { DraftDashboardSubscription, State } from "metabase/redux/store";
+import type { State } from "metabase/redux/store";
 import type { UiParameter } from "metabase-lib/v1/parameters/types";
 import type {
   Channel,
@@ -32,6 +32,7 @@ import type {
   ChannelType,
   Dashboard,
   DashboardSubscription,
+  DraftDashboardSubscription,
   ScheduleSettings,
   SubscriptionSupportingCard,
   User,
