+++ frontend/src/metabase/notifications/pulse/actions.ts
@@ -3,13 +3,13 @@ import { t } from "ttag";
 
 import { subscriptionApi } from "metabase/api";
 import { createThunkAction } from "metabase/redux";
-import type { DraftDashboardSubscription } from "metabase/redux/store";
 import { addUndo } from "metabase/redux/undo";
 import { getResponseErrorMessage } from "metabase/utils/errors";
 import type {
   ChannelApiResponse,
   CreateSubscriptionRequest,
   DashboardSubscription,
+  DraftDashboardSubscription,
   UpdateSubscriptionRequest,
 } from "metabase-types/api";
