+++ frontend/src/metabase/notifications/pulse/selectors.ts
@@ -1,5 +1,8 @@
-import type { DraftDashboardSubscription, State } from "metabase/redux/store";
-import type { ChannelApiResponse } from "metabase-types/api";
+import type { State } from "metabase/redux/store";
+import type {
+  ChannelApiResponse,
+  DraftDashboardSubscription,
+} from "metabase-types/api";
 
 export const getEditingPulse = (state: State): DraftDashboardSubscription =>
   state.pulse.editingPulse;
