+++ frontend/src/metabase/pulse/pulse.unit.spec.ts
@@ -1,6 +1,5 @@
-import type { DashboardSubscriptionData } from "metabase/redux/store";
 import MetabaseSettings from "metabase/utils/settings";
-import type { Parameter } from "metabase-types/api";
+import type { DashboardSubscriptionData, Parameter } from "metabase-types/api";
 
 import {
   type RecipientPickerValue,
